Trying to fix circular dependencies
